* Se separan los componentes en js. * Se visualiza la lista de tickets.
