+++ views/app.ejs
@@ -29,7 +29,7 @@
                     <div class="card-body">
                       <h5 class="card-title">Venta</h5>
                       <p class="card-text">Ventas</p>
-                      <a href="#" class="btn btn-primary">Ir a</a>
+                      <a href="/tickets" class="btn btn-primary">Ir a</a>
                     </div>
                 </div>
             </div>
